refactor graphic options
