+++ swmmio/defs/config.py
@@ -9,9 +9,6 @@ if os.name == 'posix':
 else:
     SWMM_ENGINE_PATH = os.path.join(ROOT_DIR, 'lib', 'windows', 'swmm5_22.exe')
 
-# path to the default geodatabase. used for some arcpy functions and parcel calcs
-GEODATABASE = r'C:\Data\ArcGIS\GDBs\LocalData.gdb'
-
 # feature class name of parcels in geodatabase
 PARCEL_FEATURES = r'PWD_PARCELS_SHEDS_PPORT'
 
@@ -19,8 +16,12 @@ PARCEL_FEATURES = r'PWD_PARCELS_SHEDS_PPORT'
 REPORT_DIR_NAME = r'Report'
 
 # path to the basemap file used to create custom basemaps
+FONT_PATH = os.path.join(ROOT_DIR, 'swmmio', 'graphics', 'fonts', 'Verdana.ttf')
+
+# path to the default geodatabase. used for some arcpy functions and parcel calcs
+GEODATABASE = r'C:\Data\ArcGIS\GDBs\LocalData.gdb'
+
+# path to the basemap file used to create custom basemaps
 BASEMAP_PATH = os.path.join(ROOT_DIR, 'swmmio', 'reporting', 'basemaps', 'index.html')
 BETTER_BASEMAP_PATH = os.path.join(ROOT_DIR, 'swmmio', 'reporting', 'basemaps', 'mapbox_base.html')
 
-# path to the basemap file used to create custom basemaps
-FONT_PATH = os.path.join(ROOT_DIR, 'swmmio', 'graphics', 'fonts', 'Verdana.ttf')
