+++ swmmio/defs/constants.py
